// src/interfaces/IPitch.ts
export default interface IPitch{
    idPitch: number;
    idTime: number;
    nomePitch: string;
    nomeTime: string;
    descricao: string;
    dataComite: string;
    situacao: number;
    situacaoDescricao: string;
    situacaoAndamento: string;
    usuarioAtualizacao: string;
    dataAtualizacao: string;
}